# frontend/src-gen/api-docs.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "CampusCoffeeConsumption API",
    "description" : "REST API for tracking the coffee consumption of SE@UHD users: users bump their own count via a secret capability link, and admins manage users and adjust counts.",
    "version" : "0.0.0"
  },
...
  "paths" : {
    "/api/users/{userId}/expenses/{expenseId}" : {
      "put" : {
        "tags" : [ "Admin expenses" ],
        "summary" : "Correct a recorded bean purchase.",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "expenseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AdminExpenseDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Admin expenses" ],
        "summary" : "Delete a recorded bean purchase.",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "expenseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}/expenses" : {
      "get" : {
        "tags" : [ "Admin expenses" ],
        "summary" : "List a user's recorded bean purchases.",
        "operationId" : "list",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ExpenseDto"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Admin expenses" ],
        "summary" : "Record a bean purchase for a user, with a kitty/private split.",
        "operationId" : "create_1",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AdminExpenseDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/expenses" : {
      "post" : {
        "tags" : [ "Expenses" ],
        "summary" : "Record the authenticated user's own bean purchase.",
        "operationId" : "record",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OwnExpenseDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummaryDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}/summary" : {
      "get" : {
        "tags" : [ "Admin accounting" ],
        "summary" : "Get a user's landing summary (admin).",
        "operationId" : "userSummary",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "exclusiveMinimum" : 0,
            "maximum" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 1000000,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummaryDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/summary" : {
      "get" : {
        "tags" : [ "Self-service" ],
        "summary" : "Get the authenticated user's landing summary.",
        "operationId" : "summary",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "exclusiveMinimum" : 0,
            "maximum" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 1000000,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummaryDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AdminExpenseDto" : {
        "type" : "object",
        "properties" : {
          "expenseType" : {
            "type" : "string",
            "enum" : [ "BEANS", "OTHER" ]
          },
          "beanName" : {
            "type" : [ "string", "null" ],
            "maxLength" : 200,
            "minLength" : 0
          },
          "weightGrams" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "maximum" : 1000,
            "minimum" : 100
          },
          "amountCents" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100000,
            "minimum" : 1
          },
          "privateAmountCents" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100000,
            "minimum" : 0
          },
          "kittyAmountCents" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100000,
            "minimum" : 0
          },
          "note" : {
            "type" : [ "string", "null" ],
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "amountCents", "expenseType", "kittyAmountCents", "privateAmountCents" ]
      },
      "ExpenseDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "buyerUserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "buyerLoginName" : {
            "type" : "string"
          },
          "expenseType" : {
            "type" : "string",
            "enum" : [ "BEANS", "OTHER" ]
          },
          "beanId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "beanName" : {
            "type" : [ "string", "null" ]
          },
          "weightGrams" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "amountCents" : {
            "type" : "integer",
            "format" : "int32"
          },
          "privateAmountCents" : {
            "type" : "integer",
            "format" : "int32"
          },
          "kittyAmountCents" : {
            "type" : "integer",
            "format" : "int32"
          },
          "note" : {
            "type" : [ "string", "null" ]
          },
          "createdAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          }
        },
        "required" : [ "amountCents", "buyerLoginName", "buyerUserId", "expenseType", "id", "kittyAmountCents", "privateAmountCents" ]
      },
...
      "ActivityEntryDto" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "CONSUMPTION", "CONSUMPTION_CANCEL", "PRIVATE_EXPENSE", "DEPOSIT", "KITTY_EXPENSE", "KITTY_ADJUSTMENT", "PRICE_CHANGE" ]
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdBy" : {
            "type" : "string"
          },
          "note" : {
            "type" : [ "string", "null" ]
          },
          "amountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "runningBalanceCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "count" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "delta" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "weightGrams" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "privateAmountCents" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          },
          "kittyAmountCents" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          }
        },
        "required" : [ "amountCents", "createdAt", "createdBy", "id", "runningBalanceCents", "type" ]
      },
      "CoffeeRatingPromptDto" : {
        "type" : "object",
        "properties" : {
          "canRate" : {
            "type" : "boolean"
          },
          "defaultBeanId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "value" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          }
        },
        "required" : [ "canRate" ]
      },
      "UserSummaryDto" : {
        "type" : "object",
        "properties" : {
          "count" : {
            "type" : "integer",
            "format" : "int32"
          },
          "priceCents" : {
            "type" : "integer",
            "format" : "int32"
          },
          "balanceCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "kittyBalanceCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cancellable" : {
            "type" : "boolean"
          },
          "summaryPanel" : {
            "type" : "string",
            "enum" : [ "BALANCE", "CUPS" ]
          },
          "firstCupAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "cupsThisWeek" : {
            "type" : "integer",
            "format" : "int32"
          },
          "cupsToday" : {
            "type" : "integer",
            "format" : "int32"
          },
          "ratingPrompt" : {
            "$ref" : "#/components/schemas/CoffeeRatingPromptDto"
          },
          "activity" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ActivityEntryDto"
            }
          }
        },
        "required" : [ "activity", "balanceCents", "cancellable", "count", "cupsThisWeek", "cupsToday", "kittyBalanceCents", "priceCents", "ratingPrompt", "summaryPanel" ]
      },
...
      "OwnExpenseDto" : {
        "type" : "object",
        "properties" : {
          "expenseType" : {
            "type" : "string",
            "enum" : [ "BEANS", "OTHER" ]
          },
          "beanName" : {
            "type" : [ "string", "null" ],
            "maxLength" : 200,
            "minLength" : 0
          },
          "weightGrams" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "maximum" : 1000,
            "minimum" : 100
          },
          "amountCents" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100000,
            "minimum" : 1
          },
          "note" : {
            "type" : [ "string", "null" ],
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "amountCents", "expenseType" ]
      },
...
      "ErrorResponse" : {
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "statusMessage" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "path" : {
            "type" : "string"
          }
        }
      }
    }
  }
}